requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
redis[asyncio]
pydantic-settings
